pyproject.toml
[tool.poetry]
name = "cfaa"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1"
newspaper3k = "^0.2.8"
pandas = "^2.2.2"
googlesearch-python = "^1.2.5"
lxml-html-clean = "^0.2.0"
python-dotenv = "^1.0.1"
openai = "^1.44.0"
django-environ = "^0.11.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
